fix pytest error
